requirements.txt
# AI Vehicle Health Assistant – Python dependencies
# Install with: pip install -r requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2

# RAG components
sentence-transformers==3.0.1
faiss-cpu==1.8.0

# LLM
anthropic==0.34.2

# Database
sqlalchemy==2.0.35

# Utilities
python-multipart==0.0.9
httpx==0.27.2
